requirements.txt
# Core RAG dependencies
faiss-cpu==1.7.4
rank-bm25==0.2.2
sentence-transformers>=5.1.1

# API and UI
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Document processing
pypdf==3.17.4
beautifulsoup4==4.12.2

# Model integration
openai>=1.10.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1
tiktoken==0.5.2
numpy==1.24.3